pyproject.toml
[project]
name = "tornado-02-api"
version = "0.1.0"
description = "Add your description here"
authors = [{name = "your name", email = "your@email.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "asyncpg>=0.30.0",
    "databases[postgresql]>=0.9.0",
    "fastapi[standard]>=0.116.1",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.35.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.10",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "google-generativeai>=0.7.2",
    "google-cloud-vision",
    "pytesseract",
    "openai",
    "beautifulsoup4",
    "lxml",
    "pypdf",
    "Pillow",
    "markdown",
    "bcrypt<4.1",
    "APScheduler>=3.10.4",
    "weasyprint>=66.0",
    "markdown2>=2.5.4",
    "pandas",
    "numpy",
    "scikit-learn",
    "pyyaml",
    "matplotlib",
    "seaborn",
    "tabulate",
    "requests",
    "pyarrow",
    "openpyxl",
]


[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.8",
    "pytest-asyncio>=0.23.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["python*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
